Move Wear OS to its own module
