+++ app/build.gradle
@@ -128,7 +128,6 @@ dependencies {
 
     // Non-free dependencies:
     playImplementation 'com.google.android.play:core:1.8.0'
-    compileOnly "com.google.android.wearable:wearable:$wearableSupportVersion"
 
     androidTestImplementation "org.awaitility:awaitility:$awaitilityVersion"
     androidTestImplementation 'com.nanohttpd:nanohttpd:2.1.1'
