+++ build.gradle
@@ -47,7 +47,7 @@ project.ext {
     exoPlayerVersion = "2.14.2"
 
     // Google Play build
-    wearableSupportVersion = "2.6.0"
+    wearableSupportVersion = "2.9.0"
 
     //Tests
     awaitilityVersion = "3.1.6"
