+++ playback/wear/README.md
@@ -1,0 +1,3 @@
+# :playback:wear
+
+This module provides Wear OS support for the Google Play version of the app.
