+++ playback/wear/src/free/java/de/danoeh/antennapod/playback/wear/WearMediaSession.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.service.playback;
+package de.danoeh.antennapod.playback.wear;
 
 import android.support.v4.media.session.MediaSessionCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
@@ -7,11 +7,11 @@ class WearMediaSession {
     /**
      * Take a custom action builder and add no extras, because this is not the Play version of the app.
      */
-    static void addWearExtrasToAction(PlaybackStateCompat.CustomAction.Builder actionBuilder) {
+    public static void addWearExtrasToAction(PlaybackStateCompat.CustomAction.Builder actionBuilder) {
         // no-op
     }
 
-    static void mediaSessionSetExtraForWear(MediaSessionCompat mediaSession) {
+    public static void mediaSessionSetExtraForWear(MediaSessionCompat mediaSession) {
         // no-op
     }
 }
